requirements.txt
yfinance
langchain
langgraph
langchain-google-genai
langchain-openai
langchain-community
finnhub-python
IPython
fastapi
uvicorn
streamlit
plotly